Consolidated all assets (images, logos, icons, SVGs) into assets.js file.
