+++ client/src/assets/assets.js
@@ -1,0 +1,273 @@
+// src/assets/images.js
+
+// SVG Imports
+import ArrowSVG from './svg/arrow.svg';
+import BackgroundElementSVG from './svg/background_element.svg';
+import ChatBubblePointerSVG from './svg/chat_bubble_pointer.svg';
+import VectorForLogo1SVG from './svg/vector_for_logo_1.svg';
+import VectorForLogo2SVG from './svg/vector_for_logo_2.svg';
+
+// Icon Imports (now referencing JS components)
+import AddIcon from './svg/icons/AddIcon';
+import BellIcon from './svg/icons/BellIcon';
+import BurgerMenuIcon from './svg/icons/BurgerMenuIcon';
+import CalendarIcon from './svg/icons/CalendarIcon';
+import CalendarOutlineIcon from './svg/icons/CalendarOutlineIcon';
+import CameraIcon from './svg/icons/CameraIcon';
+import ChatIcon from './svg/icons/ChatIcon';
+import CheckIcon from './svg/icons/CheckIcon';
+import ChevronDownIcon from './svg/icons/ChevronDownIcon';
+import ChevronLeftIcon from './svg/icons/ChevronLeftIcon';
+import ChevronRightIcon from './svg/icons/ChevronRightIcon';
+import ChevronUpIcon from './svg/icons/ChevronUpIcon';
+import ClockIcon from './svg/icons/ClockIcon';
+import CloseIcon from './svg/icons/CloseIcon';
+import DashboardIcon from './svg/icons/DashboardIcon';
+import DoubleCheckIcon from './svg/icons/DoubleCheckIcon';
+import DownloadIcon from './svg/icons/DownloadIcon';
+import EditIcon from './svg/icons/EditIcon';
+import EyeOffIcon from './svg/icons/EyeOffIcon';
+import EyeOnIcon from './svg/icons/EyeOnIcon';
+import FilterIcon from './svg/icons/FilterIcon';
+import FormIcon from './svg/icons/FormIcon';
+import GoogleIcon from './svg/icons/GoogleIcon';
+import GridIcon from './svg/icons/GridIcon';
+import HistoryIcon from './svg/icons/HistoryIcon';
+import HomeIcon from './svg/icons/HomeIcon';
+import ImageIcon from './svg/icons/ImageIcon';
+import InfoOutlineIcon from './svg/icons/InfoOutlineIcon';
+import InfoSolidIcon from './svg/icons/InfoSolidIcon';
+import InstituteIcon from './svg/icons/InstituteIcon';
+import ListIcon from './svg/icons/ListIcon';
+import MegaphoneIcon from './svg/icons/MegaphoneIcon';
+import MessengerIcon from './svg/icons/MessengerIcon';
+import MinusIcon from './svg/icons/MinusIcon';
+import MoreIcon from './svg/icons/MoreIcon';
+import PDFIcon from './svg/icons/PDFIcon';
+import PendingIcon from './svg/icons/PendingIcon';
+import PinIcon from './svg/icons/PinIcon';
+import PlusIcon from './svg/icons/PlusIcon';
+import Profile from './svg/icons/Profile';
+import ProgressIcon from './svg/icons/ProgressIcon';
+import QuestionIcon from './svg/icons/QuestionIcon';
+import ReturnIcon from './svg/icons/ReturnIcon';
+import SearchIcon from './svg/icons/SearchIcon';
+import SendIcon from './svg/icons/SendIcon';
+import SubmissionIcon from './svg/icons/SubmissionIcon';
+import TigerIcon from './svg/icons/TigerIcon';
+import TrashIcon from './svg/icons/TrashIcon';
+import UploadIcon from './svg/icons/UploadIcon';
+
+// Image Imports
+import AnnouncementSampleImage from './images/announcement_sample_image.png';
+import AppTourStep1Image from './images/app_tour_step_1_image.png';
+import AppTourStep2Image from './images/app_tour_step_2_image.png';
+import AppTourStep3Image from './images/app_tour_step_3_image.png';
+import AppTourStep4Image from './images/app_tour_step_4_image.png';
+import AppTourStep5Image from './images/app_tour_step_5_image.png';
+import AppTourThumbnail from './images/app_tour_thumbnail.png';
+import BulSULogo from './images/bulsu_logo.png';
+import CICTLogo from './images/cict_logo.png';
+import Feature1 from './images/feature_1_image.png';
+import Feature2 from './images/feature_2_image.png';
+import Feature3 from './images/feature_3_image.png';
+import Feature4 from './images/feature_4_image.png';
+import FormStatisticsHeroImage from './images/form_statistics_hero_image.png';
+import LoadingGif from './images/loading.gif';
+import LoginIllustration from './images/login_illustration.png';
+import MockupHeroImage from './images/mockup_hero_image.png';
+import OOFormThumbnail from './images/oo_form_thumbnail.png';
+import OPCRFormPDFDownloadDialogThumbnail from './images/opcr_form_pdf_download_dialog_thumbnail.png';
+import OPCRFormThumbnail from './images/opcr_form_thumbnail.png';
+import OPCRFormTrackerHeroImage from './images/opcr_form_tracker_hero_image.png';
+import OPCRLogo from './images/opcr_logo.png';
+
+// Logos Imports
+import AssetManagementUnit from './images/logos/asset_management_unit.png';
+import BulacanStateUniversityLogo from './images/logos/bulacan_state_university.png';
+import BulacanStateUniversityBustosCampusLogo from './images/logos/bulacan_state_university_bustos_campus.png';
+import BulacanStateUniversityHagonoyCampusLogo from './images/logos/bulacan_state_university_hagonoy_campus.png';
+import BulacanStateUniversityMenesesCampusLogo from './images/logos/bulacan_state_university_meneses_campus.png';
+import BulacanStateUniversitySanRafaelCampusLogo from './images/logos/bulacan_state_university_san_rafael_campus.png';
+import BulacanStateUniversitySarmientoCampusLogo from './images/logos/bulacan_state_university_sarmiento_campus.png';
+import CareerDevelopmentInternshipLogo from './images/logos/career_development_internship.png';
+import CollegeOfArchitectureAndFineArts from './images/logos/college_of_architecture_and_fine_arts.png';
+import CollegeOfArtsAndLetters from './images/logos/college_of_arts_and_letters.png';
+import CollegeOfBusinessAdministration from './images/logos/college_of_business_administration.png';
+import CollegeOfCriminalJusticeEducation from './images/logos/college_of_criminal_justice_education.png';
+import CollegeOfEducation from './images/logos/college_of_education.png';
+import CollegeOfEngineering from './images/logos/college_of_engineering.png';
+import CollegeOfHospitalityAndTourismManagement from './images/logos/college_of_hospitality_and_tourism_management.png';
+import CollegeOfIndustrialTechnology from './images/logos/college_of_industrial_technology.png';
+import CollegeOfInformationAndCommunicationsTechnology from './images/logos/college_of_information_and_communications_technology.png';
+import CollegeOfLaw from './images/logos/college_of_law.png';
+import CollegeOfNursing from './images/logos/college_of_nursing.png';
+import CollegeOfScience from './images/logos/college_of_science.png';
+import ExternalRelationsAndInternationalization from './images/logos/external_relations_and_internationalization.png';
+import GraduateSchool from './images/logos/graduate_school.png';
+import GuidanceAndCounselingServiceCenter from './images/logos/guidance_and_counseling_service_center.png';
+import HumanResourceManagementOffice from './images/logos/human_resource_management_office.png';
+import LaboratoryHighSchool from './images/logos/laboratory_high_school.png';
+import NationalServiceTrainingProgram from './images/logos/national_service_training_program.png';
+import OfficeOfInternationalAffairs from './images/logos/office_of_international_affairs.png';
+import OfficeOfTheCampusDirectorForExtensionServices from './images/logos/office_of_the_campus_director_for_extension_services.png';
+import OfficeOfTheCampusDirectorForResearchAndInnovation from './images/logos/office_of_the_campus_director_for_research_and_innovation.png';
+import OfficeOfTheChancellorExternalCampus from './images/logos/office_of_the_chancellor_external_campus.png';
+import OfficeOfTheCultureAndArts from './images/logos/office_of_the_culture_and_arts.png';
+import OfficeOfTheScholarships from './images/logos/office_of_the_scholarships.png';
+import OfficeOfTheStudentGovernment from './images/logos/office_of_the_student_government.png';
+import OfficeOfTheStudentOrganizationActivities from './images/logos/office_of_the_student_organization_activities.png';
+import OfficeOfTheViceChancellorForStudentAffairsMainCampus from './images/logos/office_of_the_vice_chancellor_for_student_affairs_main_campus.png';
+import OfficeOfTheVicePresidentForResearchAndDevelopmentExtension from './images/logos/office_of_the_vice_president_for_research_and_development_and_extension.png';
+import PlacementAndEmploymentServicesOffice from './images/logos/placement_and_employment_services_office.png';
+import PlanningAndDevelopmentOffice from './images/logos/planning_and_development_office.png';
+import SportsDevelopmentOffice from './images/logos/sports_development_office.png';
+import StudentPublicationOffice from './images/logos/student_publication_office.png';
+import StudentWelfareAndDevelopmentOffice from './images/logos/student_welfare_and_development_office.png';
+import UniversityInfirmaryUnit from './images/logos/university_infirmary_unit.png';
+import UniversityLibrary from './images/logos/university_library.png';
+
+// SVG Illustrations Imports
+import AppTourCompletionIllustration from './svg/illustrations/app_tour_completion_illustration.svg';
+import ChangedPasswordSuccessfully from './svg/illustrations/changed_password_successfully.svg';
+import EmailAddressVerifiedIllustration from './svg/illustrations/email_address_verified_illustration.svg';
+import EmptyForm from './svg/illustrations/empty_form.svg';
+import EmptyFormsIllustration from './svg/illustrations/empty_forms_illustration.svg';
+import FormSubmittedSuccessfully from './svg/illustrations/form_submitted_successfully.svg';
+import MaintenanceIllustration from './svg/illustrations/maintenance_illustration.svg';
+import OnboardIllustration from './svg/illustrations/onboard.svg';
+import PasswordResetIllustration from './svg/illustrations/password_reset_illustration.svg';
+import RegistrationSuccessfullyIllustration from './svg/illustrations/registration_successfully.svg';
+
+// Export all images, logos, icons, and illustrations
+export {
+  ArrowSVG,
+  BackgroundElementSVG,
+  ChatBubblePointerSVG,
+  VectorForLogo1SVG,
+  VectorForLogo2SVG,
+  AddIcon,
+  BellIcon,
+  BurgerMenuIcon,
+  CalendarIcon,
+  CalendarOutlineIcon,
+  CameraIcon,
+  ChatIcon,
+  CheckIcon,
+  ChevronDownIcon,
+  ChevronLeftIcon,
+  ChevronRightIcon,
+  ChevronUpIcon,
+  ClockIcon,
+  CloseIcon,
+  DashboardIcon,
+  DoubleCheckIcon,
+  DownloadIcon,
+  EditIcon,
+  EyeOffIcon,
+  EyeOnIcon,
+  FilterIcon,
+  FormIcon,
+  GoogleIcon,
+  GridIcon,
+  HistoryIcon,
+  HomeIcon,
+  ImageIcon,
+  InfoOutlineIcon,
+  InfoSolidIcon,
+  InstituteIcon,
+  ListIcon,
+  MegaphoneIcon,
+  MessengerIcon,
+  MinusIcon,
+  MoreIcon,
+  PDFIcon,
+  PendingIcon,
+  PinIcon,
+  PlusIcon,
+  Profile,
+  ProgressIcon,
+  QuestionIcon,
+  ReturnIcon,
+  SearchIcon,
+  SendIcon,
+  SubmissionIcon,
+  TigerIcon,
+  TrashIcon,
+  UploadIcon,
+  AnnouncementSampleImage,
+  AppTourStep1Image,
+  AppTourStep2Image,
+  AppTourStep3Image,
+  AppTourStep4Image,
+  AppTourStep5Image,
+  AppTourThumbnail,
+  BulSULogo,
+  CICTLogo,
+  Feature1,
+  Feature2,
+  Feature3,
+  Feature4,
+  FormStatisticsHeroImage,
+  LoadingGif,
+  LoginIllustration,
+  MockupHeroImage,
+  OOFormThumbnail,
+  OPCRFormPDFDownloadDialogThumbnail,
+  OPCRFormThumbnail,
+  OPCRFormTrackerHeroImage,
+  OPCRLogo,
+  AssetManagementUnit,
+  BulacanStateUniversityLogo,
+  BulacanStateUniversityBustosCampusLogo,
+  BulacanStateUniversityHagonoyCampusLogo,
+  BulacanStateUniversityMenesesCampusLogo,
+  BulacanStateUniversitySanRafaelCampusLogo,
+  BulacanStateUniversitySarmientoCampusLogo,
+  CareerDevelopmentInternshipLogo,
+  CollegeOfArchitectureAndFineArts,
+  CollegeOfArtsAndLetters,
+  CollegeOfBusinessAdministration,
+  CollegeOfCriminalJusticeEducation,
+  CollegeOfEducation,
+  CollegeOfEngineering,
+  CollegeOfHospitalityAndTourismManagement,
+  CollegeOfIndustrialTechnology,
+  CollegeOfInformationAndCommunicationsTechnology,
+  CollegeOfLaw,
+  CollegeOfNursing,
+  CollegeOfScience,
+  ExternalRelationsAndInternationalization,
+  GraduateSchool,
+  GuidanceAndCounselingServiceCenter,
+  HumanResourceManagementOffice,
+  LaboratoryHighSchool,
+  NationalServiceTrainingProgram,
+  OfficeOfInternationalAffairs,
+  OfficeOfTheCampusDirectorForExtensionServices,
+  OfficeOfTheCampusDirectorForResearchAndInnovation,
+  OfficeOfTheChancellorExternalCampus,
+  OfficeOfTheCultureAndArts,
+  OfficeOfTheScholarships,
+  OfficeOfTheStudentGovernment,
+  OfficeOfTheStudentOrganizationActivities,
+  OfficeOfTheViceChancellorForStudentAffairsMainCampus,
+  OfficeOfTheVicePresidentForResearchAndDevelopmentExtension,
+  PlacementAndEmploymentServicesOffice,
+  PlanningAndDevelopmentOffice,
+  SportsDevelopmentOffice,
+  StudentPublicationOffice,
+  StudentWelfareAndDevelopmentOffice,
+  UniversityInfirmaryUnit,
+  UniversityLibrary,
+  AppTourCompletionIllustration,
+  ChangedPasswordSuccessfully,
+  EmailAddressVerifiedIllustration,
+  EmptyForm,
+  EmptyFormsIllustration,
+  FormSubmittedSuccessfully,
+  MaintenanceIllustration,
+  OnboardIllustration,
+  PasswordResetIllustration,
+  RegistrationSuccessfullyIllustration
+};
