--- client/src/assets/images.js
@@ -1,111 +1,0 @@
-// src/assets/images.js
-
-// SVG Imports
-import ArrowSVG from './svg/arrow.svg';
-
-// Image Imports
-import BulSULogo from './images/bulsu_logo.png';
-import OPCRLogo from './images/opcr_logo.png';
-import MockupHeroImage from './images/mockup_hero_image.png';
-import OPCRFormTrackerHeroImage from './images/opcr_form_tracker_hero_image.png';
-import FormStatisticsHeroImage from './images/form_statistics_hero_image.png';
-import Feature1 from './images/feature_1_image.png';
-import Feature2 from './images/feature_2_image.png';
-import Feature3 from './images/feature_3_image.png';
-import Feature4 from './images/feature_4_image.png';
-
-// Logos Imports
-import AssetManagementUnit from './images/logos/asset_management_unit.png';
-import BulacanStateUniversityLogo from './images/logos/bulacan_state_university.png';
-import BulacanStateUniversityHagonoyCampusLogo from './images/logos/bulacan_state_university_hagonoy_campus.png';
-import BulacanStateUniversityMenesesCampusLogo from './images/logos/bulacan_state_university_meneses_campus.png';
-import BulacanStateUniversitySarmientoCampusLogo from './images/logos/bulacan_state_university_sarmiento_campus.png';
-import CollegeOfArchitectureAndFineArts from './images/logos/college_of_architecture_and_fine_arts.png';
-import CollegeOfArtsAndLetters from './images/logos/college_of_arts_and_letters.png';
-import CollegeOfBusinessAdministration from './images/logos/college_of_business_administration.png';
-import CollegeOfCriminalJusticeEducation from './images/logos/college_of_criminal_justice_education.png';
-import CollegeOfEducation from './images/logos/college_of_education.png';
-import CollegeOfEngineering from './images/logos/college_of_engineering.png';
-import CollegeOfHospitalityAndTourismManagement from './images/logos/college_of_hospitality_and_tourism_management.png';
-import CollegeOfIndustrialTechnology from './images/logos/college_of_industrial_technology.png';
-import CollegeOfInformationAndCommunicationsTechnology from './images/logos/college_of_information_and_communications_technology.png';
-import CollegeOfLaw from './images/logos/college_of_law.png';
-import CollegeOfNursing from './images/logos/college_of_nursing.png';
-import CollegeOfScience from './images/logos/college_of_science.png';
-import ExternalRelationsAndInternationalization from './images/logos/external_relations_and_internationalization.png';
-import GraduateSchool from './images/logos/graduate_school.png';
-import GuidanceAndCounselingServiceCenter from './images/logos/guidance_and_counseling_service_center.png';
-import HumanResourceManagementOffice from './images/logos/human_resource_management_office.png';
-import LaboratoryHighSchool from './images/logos/laboratory_high_school.png';
-import NationalServiceTrainingProgram from './images/logos/national_service_training_program.png';
-import OfficeOfInternationalAffairs from './images/logos/office_of_international_affairs.png';
-import OfficeOfTheCampusDirectorForExtensionServices from './images/logos/office_of_the_campus_director_for_extension_services.png';
-import OfficeOfTheCampusDirectorForResearchAndInnovation from './images/logos/office_of_the_campus_director_for_research_and_innovation.png';
-import OfficeOfTheChancellorExternalCampus from './images/logos/office_of_the_chancellor_external_campus.png';
-import OfficeOfTheCultureAndArts from './images/logos/office_of_the_culture_and_arts.png';
-import OfficeOfTheScholarships from './images/logos/office_of_the_scholarships.png';
-import OfficeOfTheStudentGovernment from './images/logos/office_of_the_student_government.png';
-import OfficeOfTheStudentOrganizationActivities from './images/logos/office_of_the_student_organization_activities.png';
-import OfficeOfTheViceChancellorForStudentAffairsMainCampus from './images/logos/office_of_the_vice_chancellor_for_student_affairs_main_campus.png';
-import OfficeOfTheVicePresidentForResearchAndDevelopmentExtension from './images/logos/office_of_the_vice_president_for_research_and_development_and_extension.png';
-import PlacementAndEmploymentServicesOffice from './images/logos/placement_and_employment_services_office.png';
-import PlanningAndDevelopmentOffice from './images/logos/planning_and_development_office.png';
-import SportsDevelopmentOffice from './images/logos/sports_development_office.png';
-import StudentPublicationOffice from './images/logos/student_publication_office.png';
-import StudentWelfareAndDevelopmentOffice from './images/logos/student_welfare_and_development_office.png';
-import UniversityInfirmaryUnit from './images/logos/university_infirmary_unit.png';
-import UniversityLibrary from './images/logos/university_library.png';
-
-// Export all images and logos
-export {
-  ArrowSVG,
-  BulSULogo,
-  OPCRLogo,
-  MockupHeroImage,
-  OPCRFormTrackerHeroImage,
-  FormStatisticsHeroImage,
-  Feature1,
-  Feature2,
-  Feature3,
-  Feature4,
-  AssetManagementUnit,
-  BulacanStateUniversityLogo,
-  BulacanStateUniversityHagonoyCampusLogo,
-  BulacanStateUniversityMenesesCampusLogo,
-  BulacanStateUniversitySarmientoCampusLogo,
-  CollegeOfArchitectureAndFineArts,
-  CollegeOfArtsAndLetters,
-  CollegeOfBusinessAdministration,
-  CollegeOfCriminalJusticeEducation,
-  CollegeOfEducation,
-  CollegeOfEngineering,
-  CollegeOfHospitalityAndTourismManagement,
-  CollegeOfIndustrialTechnology,
-  CollegeOfInformationAndCommunicationsTechnology,
-  CollegeOfLaw,
-  CollegeOfNursing,
-  CollegeOfScience,
-  ExternalRelationsAndInternationalization,
-  GraduateSchool,
-  GuidanceAndCounselingServiceCenter,
-  HumanResourceManagementOffice,
-  LaboratoryHighSchool,
-  NationalServiceTrainingProgram,
-  OfficeOfInternationalAffairs,
-  OfficeOfTheCampusDirectorForExtensionServices,
-  OfficeOfTheCampusDirectorForResearchAndInnovation,
-  OfficeOfTheChancellorExternalCampus,
-  OfficeOfTheCultureAndArts,
-  OfficeOfTheScholarships,
-  OfficeOfTheStudentGovernment,
-  OfficeOfTheStudentOrganizationActivities,
-  OfficeOfTheViceChancellorForStudentAffairsMainCampus,
-  OfficeOfTheVicePresidentForResearchAndDevelopmentExtension,
-  PlacementAndEmploymentServicesOffice,
-  PlanningAndDevelopmentOffice,
-  SportsDevelopmentOffice,
-  StudentPublicationOffice,
-  StudentWelfareAndDevelopmentOffice,
-  UniversityInfirmaryUnit,
-  UniversityLibrary
-};
