--- client/src/assets/svg/index.js
@@ -1,101 +1,0 @@
-import AddIcon from './icons/AddIcon'
-import BellIcon from './icons/BellIcon'
-import BurgerMenuIcon from './icons/BurgerMenuIcon'
-import CalendarIcon from './icons/CalendarIcon'
-import CalendarOutlineIcon from './icons/CalendarOutlineIcon'
-import CameraIcon from './icons/CameraIcon'
-import ChatIcon from './icons/ChatIcon'
-import CheckIcon from './icons/CheckIcon'
-import ChevronDownIcon from './icons/ChevronDownIcon'
-import ChevronLeftIcon from './icons/ChevronLeftIcon'
-import ChevronRightIcon from './icons/ChevronRightIcon'
-import ChevronUpIcon from './icons/ChevronUpIcon'
-import ClockIcon from './icons/ClockIcon'
-import CloseIcon from './icons/CloseIcon'
-import DashboardIcon from './icons/DashboardIcon'
-import DoubleCheckIcon from './icons/DoubleCheckIcon'
-import DownloadIcon from './icons/DownloadIcon'
-import EditIcon from './icons/EditIcon'
-import EyeOffIcon from './icons/EyeOffIcon'
-import EyeOnIcon from './icons/EyeOnIcon'
-import FilterIcon from './icons/FilterIcon'
-import FormIcon from './icons/FormIcon'
-import GoogleIcon from './icons/GoogleIcon'
-import GridIcon from './icons/GridIcon'
-import HistoryIcon from './icons/HistoryIcon'
-import HomeIcon from './icons/HomeIcon'
-import ImageIcon from './icons/ImageIcon'
-import InfoOutlineIcon from './icons/InfoOutlineIcon'
-import InfoSolidIcon from './icons/InfoSolidIcon'
-import InstituteIcon from './icons/InstituteIcon'
-import ListIcon from './icons/ListIcon'
-import MegaphoneIcon from './icons/MegaphoneIcon'
-import MessengerIcon from './icons/MessengerIcon'
-import MinusIcon from './icons/MinusIcon'
-import MoreIcon from './icons/MoreIcon'
-import PDFIcon from './icons/PDFIcon'
-import PendingIcon from './icons/PendingIcon'
-import PinIcon from './icons/PinIcon'
-import PlusIcon from './icons/PlusIcon'
-import Profile from './icons/Profile'
-import ProgressIcon from './icons/ProgressIcon'
-import QuestionIcon from './icons/QuestionIcon'
-import ReturnIcon from './icons/ReturnIcon'
-import SearchIcon from './icons/SearchIcon'
-import SendIcon from './icons/SendIcon'
-import Submissionicon from './icons/SubmissionIcon'
-import TigerIcon from './icons/TigerIcon'
-import TrashIcon from './icons/TrashIcon'
-import UploadIcon from './icons/UploadIcon'
-
-export {
-  AddIcon,
-  BellIcon,
-  BurgerMenuIcon,
-  CalendarIcon,
-  CalendarOutlineIcon,
-  CameraIcon,
-  ChatIcon,
-  CheckIcon,
-  ChevronDownIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  ChevronUpIcon,
-  ClockIcon,
-  CloseIcon,
-  DashboardIcon,
-  DoubleCheckIcon,
-  DownloadIcon,
-  EditIcon,
-  EyeOffIcon,
-  EyeOnIcon,
-  FilterIcon,
-  FormIcon,
-  GoogleIcon,
-  GridIcon,
-  HistoryIcon,
-  HomeIcon,
-  ListIcon,
-  ImageIcon,
-  InfoOutlineIcon,
-  InfoSolidIcon,
-  InstituteIcon,
-  MegaphoneIcon,
-  MessengerIcon,
-  MinusIcon,
-  MoreIcon,
-  PDFIcon,
-  PendingIcon,
-  PinIcon,
-  PlusIcon,
-  Profile,
-  ProgressIcon,
-  QuestionIcon,
-  ReturnIcon,
-  SearchIcon,
-  SendIcon,
-  Submissionicon,
-  TigerIcon,
-  TrashIcon,
-  UploadIcon
-}
