+++ client/src/pages/LandingPage.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Header, Hero, Features, FAQs, Footer } from '../components/LandingPage/index';
-import { PlanningAndDevelopmentOffice } from '../assets/images';
+import { PlanningAndDevelopmentOffice } from '../assets/assets';
 import '../styles/landing_page.css'
 
 
@@ -13,7 +13,7 @@ const HomePage = () => {
     <section id="landing_page">
 
 <div id="intro">
-        <img src={PlanningAndDevelopmentOffice} alt="Logo" />
+        <img src={ PlanningAndDevelopmentOffice } alt="Logo" />
         <div className="rectangle"></div>
       </div>
       <div className="container">
